requirements.txt
# Dependencies shared across backend services for local testing
fastapi==0.119.0
uvicorn==0.37.0
pytest==8.4.2
pytest-cov==7.0.0
sqlmodel==0.0.27
python-dotenv==1.1.1
pydantic-settings==2.11.0
python-multipart==0.0.20
psycopg2-binary==2.9.11
httpx==0.28.1
pydantic[email]==2.12.2
email-validator==2.3.0
Faker==37.11.0
python-jose==3.5.0
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
passlib==1.7.4
